src/requirements.txt
APScheduler==3.5.3
atomicwrites==1.3.0
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
more-itertools==5.0.0
pep8==1.7.1
pluggy==0.8.1
py==1.7.0
pycodestyle==2.5.0
pytest==4.2.0
pytest-django==3.4.7
pytz==2018.9
requests==2.21.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
